% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andamento.R
\name{andamento}
\alias{andamento}
\title{Andamento nel tempo del numero di tweet, favoriti e retweet}
\usage{
andamento(twtbl, freq, grafico = TRUE)
}
\arguments{
\item{twtbl}{Insieme di tweet.}

\item{freq}{Ogni quanto vengono raggruppati i tweet.}

\item{grafico}{Operatore logico.}
}
\value{
Un grafico o una tibble riassuntiva degli andamenti del numero di
 tweet, favoriti e retweet nel tempo.
}
\description{
Osserva l'andamento nel tempo del numero di tweet, favoriti e retweet in un
determinato periodo di tempo, da specificare in freq.
Se grafico = TRUE, restituisce il grafico dei tre andamenti, se grafico
= FALSE, restituisce solo la tibble riassuntiva.
}
\examples{
\dontrun{

tw_amala <- search_tweets(
q = "inter", n = 5000, type = "recent", include_rts = T, lang="it"
)

andamento(tw_amala, "38 mins", grafico = F)
andamento(tw_amala, "2 hours", grafico = T)
}
}
