% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTxt.R
\name{cleanTxt}
\alias{cleanTxt}
\title{Funzione per la pulizia dei testi}
\usage{
cleanTxt(x = NULL, conv = FALSE, hash = TRUE, hashlist = NULL,
  mention = TRUE, mentlist = NULL, numeri = FALSE, minusc = TRUE)
}
\arguments{
\item{x}{Vettore dei testi (tweets).}

\item{conv}{Logico, se \code{TRUE} converte da UTF-8 a latin1.}

\item{hash}{Logico, se \code{TRUE} rimuove per intero gli hashtags.}

\item{hashlist}{Lista con gli hashtags, necessaria se \code{hash = FALSE}
(è la lista \emph{hashtag} nell’oggetto tweet, restituito da
\code{get_timeline} o da \code{search_tweet}).}

\item{mention}{Logico, se \code{TRUE} rimuove per intero le mentions.}

\item{mentlist}{Lista con le mentions, necessaria se \code{mention = FALSE}
(\code{oggetto.tweet$mentions_screen_name}).}

\item{numeri}{Logico, se \code{TRUE} rimuove tutti i numeri.}

\item{minusc}{Logico, se \code{TRUE} trasforma i testi in minuscolo.}
}
\value{
Il vettore dei testi pulito, più semplice da utilizzare per le analisi.
}
\description{
Questa funzione facilita le operazioni di pre-trattamento e pulizia dei
testi, permettendo di personalizzare le operazioni che si vogliono eseguire.
Di default:
\itemize{
\item \bold{non} viene effettuata la conversione da \emph{utf-8} a \emph{latin1}
\item vengono rimossi gli \emph{hashtag} e le \emph{mention}
\item \bold{non} vengono rimossi i numeri
\item viene effettuata la conversione a minuscolo.
}
}
\examples{
\dontrun{

tw_amala <- search_tweets(
q = "inter", n = 5000, type = "recent", include_rts = T, lang="it"
)

cleanTxt(
tw_amala$text,
conv = TRUE,
hash = TRUE,
mention = TRUE,
numeri = TRUE
minusc = TRUE)
}
}
