% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemmaTDM.R
\name{lemmaTDM}
\alias{lemmaTDM}
\title{Funzione per la lemmatizzazione di una Term-Document Matrix}
\usage{
lemmaTDM(x = NULL)
}
\arguments{
\item{x}{Data.frame con Term-Document Matrix.}
}
\value{
Un data frame LemmiDocumenti, con la categoria grammaticale e la
frequenza totale dei lemmi.
}
\description{
Questa funzione richiede in input un data.frame con una matrice
TerminiDocumenti, come quella prodotta dalla funzione \code{doWrdCld}.
Questa funzione necessita dell'utilizzo del vocabolario \code{formelemmi},
già presente di default nella libreria.
}
\examples{
\dontrun{

tw_amala <- search_tweets(
q = "inter", n = 5000, type = "recent", include_rts = T, lang="it"
)

tw_amala$clean <- cleanTxt(tw_amala$text)

tw_amala_tdm <- doWrdCld(tw_amala$clean, plot = FALSE)

tw_amala_lem <- lemmaTDM(tw_amala_tdm)

}
}
