% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Funzione interattiva per l'autenticazione via access token}
\usage{
login(insicuro = FALSE)
}
\arguments{
\item{insicuro}{Operatore logico.}
}
\description{
Segui le istruzioni e compila con le informazioni che trovi sul sito
developer.twitter.com dopo aver creato la tua app twitter secondo le
indicazioni che si trovano sulle dispense del Corso.
Inserisci insicuro = TRUE per un check dei dati inseriti.
}
