% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visNGram.R
\name{visNGram}
\alias{visNGram}
\title{Funzione per visualizzare n-grammi}
\usage{
visNGram(x = NULL, ngrI = 2, ngrF = 3, nn = 20)
}
\arguments{
\item{x}{Vettore dei testi "puliti".}

\item{ngrI}{Valore minimo lunghezza n-grammi (default = 2).}

\item{ngrF}{Valore massimo lunghezza n-grammi (default = 3).}

\item{nn}{Numero di n-grammi da visualizzare per ciascuna lunghezza.}
}
\value{
Gli n-grammi di lunghezza definita dai parametri \code{ngrI} e \code{ngrF}
}
\description{
Questa funzione facilita le operazioni di visualizzazione degli n-grammi,
permettendo di personalizzare le operazioni che si vogliono eseguire.
Una volta visualizzati gli n-grammi delle lunghezze specificate è necessario
specificare manualmente le sostituzioni da effettuare,
utilizzando la funzione \code{gsub()}
}
\examples{
\dontrun{

tw_amala <- search_tweets(
q = "inter", n = 5000, type = "recent", include_rts = T, lang="it"
)

tw_amala$clean <- cleanTxt(tw_amala$text)

visNGram(tw_amala$clean, ngrI = 2, ngrF = 4, nn = 10)

}
}
